feat(home): create main page structure with Layout component
